fileUploadsWebJob/requirements.txt
azure-identity==1.13.0
quart==0.18.4
langchain==0.1.2
langchain_experimental==0.0.49
openai[datalib]==0.27.8
tiktoken==0.4.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.14.1
uvicorn[standard]==0.23.2
aiohttp==3.8.5
azure-monitor-opentelemetry==1.0.0b15
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-aiohttp-client==0.40b0
Werkzeug==2.2.2
aioodbc
faiss-cpu==1.7.4
pypdf==3.17.1
unstructured==0.11.0
python-pptx==0.6.23
docx2txt==0.8
wheel==0.42.0
networkx==3.2.1
SQLAlchemy==1.4.50
PyMuPDF==1.24.6
Spire.Doc==12.6.0
docx2pdf==0.1.8
azure-ai-translation-document==1.0.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-speech == 1.40.0
nltk==3.8.1
office365
office365-REST-Python-Client
python-docx
PyPDF2
azure-storage-queue
